Update package dependencies and refactor components for improved functionality
- Added @mui/x-date-pickers and date-fns to package dependencies for enhanced date handling.
- Replaced Plan and Reality components with a new Log component in App for streamlined navigation.
- Updated Dashboard to include quick action buttons for logging meals and viewing logs.
- Adjusted navigation logic in Layout to reflect the new structure and removed references to deleted components.
- Enhanced mock data to include log entries for better tracking of meals and activities.
