+++ CHANGES_SUMMARY.md
@@ -1,0 +1,97 @@
+# Changes Summary: Plan & Reality → Log Unification
+
+## Files Modified
+
+### 1. `src/components/Log.tsx` (NEW)
+- **Created**: New unified Log component that combines planning and logging functionality
+- **Features**:
+  - Date-based navigation (past, current, future)
+  - Context-aware UI (planning for future, logging for past/current)
+  - Unified meal management (planned vs. actual)
+  - Activity logging (past/current only)
+  - Smart floating action buttons based on date context
+
+### 2. `src/App.tsx`
+- **Removed**: Imports for `Plan` and `Reality` components
+- **Added**: Import for new `Log` component
+- **Updated**: Route from `/plan` and `/reality` to `/log`
+
+### 3. `src/components/Layout.tsx`
+- **Removed**: `Settings` icon import (no longer needed)
+- **Updated**: Navigation items from 5 to 4 items
+- **Changed**: "Plan" and "Reality" → "Log"
+- **Updated**: Navigation routing to use `/log`
+
+### 4. `src/components/Dashboard.tsx`
+- **Added**: `Button` import from Material-UI
+- **Added**: Quick Actions section with links to Log functionality
+- **Enhanced**: Better context for daily entries display
+
+### 5. `src/types/index.ts`
+- **Added**: `LogEntry` interface for unified data representation
+- **Updated**: `DailyProgress` interface to include `logEntries` array
+
+### 6. `src/data/mockData.ts`
+- **Added**: `logEntries` property to `mockDailyProgress` with sample data
+
+## Files Deleted
+
+### 1. `src/components/Plan.tsx`
+- **Removed**: Entire Plan component (functionality merged into Log)
+
+### 2. `src/components/Reality.tsx`
+- **Removed**: Entire Reality component (functionality merged into Log)
+
+## Dependencies Added
+
+### 1. `@mui/x-date-pickers`
+- **Added**: For date picker functionality in Log component
+- **Added**: `date-fns` as peer dependency
+
+## Key Benefits of the Changes
+
+1. **Simplified Navigation**: Single "Log" item instead of separate "Plan" and "Reality"
+2. **Better User Experience**: Context-aware interface based on selected date
+3. **Unified Data Model**: Single view for all meal and activity data
+4. **Improved Workflow**: Natural progression from planning to logging
+5. **Reduced Cognitive Load**: Users don't need to switch between separate views
+
+## Technical Improvements
+
+1. **Type Safety**: Enhanced TypeScript interfaces with `LogEntry` type
+2. **Component Reusability**: Single component handles multiple use cases
+3. **Date Handling**: Proper date picker integration with Material-UI
+4. **Responsive Design**: Maintains mobile-first approach
+5. **Backward Compatibility**: Existing data models preserved
+
+## User Workflow Changes
+
+### Before (Separate Views)
+- User navigates to "Plan" to set future meals
+- User navigates to "Reality" to log actual consumption
+- Need to switch between views to compare plans vs. reality
+
+### After (Unified Log)
+- User navigates to "Log" and selects appropriate date
+- Future dates: Show planning interface
+- Past/Current dates: Show logging interface
+- All data visible in single, contextual view
+
+## Testing Status
+
+- ✅ TypeScript compilation successful
+- ✅ All imports resolved
+- ✅ Navigation updated
+- ✅ Routing configured
+- ✅ Mock data updated
+- ✅ Component structure complete
+- ✅ Development server running successfully
+- ✅ No compilation errors in browser
+
+## Next Steps
+
+1. **User Testing**: Test the new unified Log interface
+2. **Performance**: Monitor component performance with larger datasets
+3. **Accessibility**: Ensure date picker meets accessibility standards
+4. **Mobile Optimization**: Test mobile experience thoroughly
+5. **Data Migration**: Plan for any existing user data migration needs
