+++ LOG_UNIFICATION.md
@@ -1,0 +1,92 @@
+# Log Unification
+
+## Overview
+The Digital Nutritionist app has been updated to unify the "Plan" and "Reality" concepts into a single "Log" system. This provides a more intuitive and streamlined user experience for managing nutrition and activity data.
+
+## Changes Made
+
+### 1. New Log Component (`src/components/Log.tsx`)
+- **Unified Interface**: Combines planning and logging functionality in one view
+- **Date Navigation**: Users can navigate between past, current, and future dates
+- **Context-Aware Actions**: 
+  - Future dates: Show planning options (Plan Meal)
+  - Past/Current dates: Show logging options (Log Meal, Log Activity)
+- **Smart Display**: Shows appropriate data based on date context
+- **Unified Data Model**: Handles both planned and actual meals/activities seamlessly
+
+### 2. Updated Navigation (`src/components/Layout.tsx`)
+- Replaced separate "Plan" and "Reality" navigation items with single "Log" item
+- Updated routing to use `/log` instead of `/plan` and `/reality`
+- Simplified bottom navigation from 5 to 4 items
+
+### 3. Updated App Routing (`src/App.tsx`)
+- Replaced Plan and Reality route imports with Log component
+- Updated route paths to use `/log`
+
+### 4. Enhanced Dashboard (`src/components/Dashboard.tsx`)
+- Added Quick Actions section with links to Log functionality
+- Improved context for daily entries display
+- Better integration with the unified Log system
+
+### 5. Enhanced Types (`src/types/index.ts`)
+- Added `LogEntry` interface for unified data representation
+- Updated `DailyProgress` to include log entries
+
+## Key Features
+
+### Date-Based Context
+- **Future Dates**: Focus on planning meals and activities
+- **Current Date**: Balance of planning and logging
+- **Past Dates**: Focus on logging what actually happened
+
+### Unified Meal Management
+- Single interface for both planned and actual meals
+- Visual distinction between planned (blue) and logged (orange) items
+- Seamless conversion from planned to actual
+
+### Activity Logging
+- Activities can only be logged for past/current dates (not planned for future)
+- Consistent with real-world usage patterns
+
+### Smart UI Adaptation
+- Floating action buttons change based on date context
+- Summary cards adapt to show relevant information
+- Tab content adjusts based on what's appropriate for the selected date
+
+## Benefits
+
+1. **Simplified Navigation**: Users don't need to switch between separate views
+2. **Better Context**: Clear understanding of what can be done on different dates
+3. **Improved Workflow**: Natural progression from planning to logging
+4. **Reduced Cognitive Load**: Single interface for related functionality
+5. **Better Data Relationships**: Easier to see connections between plans and reality
+
+## Usage
+
+### Planning Meals (Future Dates)
+1. Navigate to a future date in the Log
+2. Use "Plan Meal" button to add planned meals
+3. Set meal details including time and calories
+
+### Logging Meals (Past/Current Dates)
+1. Navigate to past or current date
+2. Use "Log Meal" button to record what was actually consumed
+3. Add notes about the meal if needed
+
+### Logging Activities
+1. Navigate to past or current date
+2. Use "Log Activity" button to record exercise
+3. Include duration and calories burned
+
+### Viewing History
+- Navigate to any past date to see what was logged
+- Compare planned vs. actual consumption
+- Track progress over time
+
+## Technical Implementation
+
+- Uses Material-UI components for consistent design
+- Implements date picker for easy navigation
+- Responsive design for mobile and desktop
+- Type-safe implementation with TypeScript
+- Maintains existing data models for backward compatibility
